refactor: attestation.rs and client rename
